Update project.cpp
the MinGW version of dirent is simplified. the d_type field is no longer available. so if you compile using MinGW it gives that error.
I use the "d_type" field to know the type (file, folder) of an entry, 
as a workaround, instead of testing like:
  if (Dirent-> d_type! = DT_DIR)
{
     // add the file
}
we can test as follow:
DIR * TEMP;
if ((TEMP = opendir (Dirent-> d_name)))
{
                 closedir (TEMP);
}
else
{
         // add the file
}